Set z3 answer to 'unknown' if result is NEQ and query contains uninterpreted functions
Overall there are 4 different cases that might occur:

1. the result is EQ and the query does NOT contain UIFs,
2. the result is EQ and the query DOES contain UIFs,
3. the result is NEQ and the query does NOT contain UIFs,
4. the result is NEQ and the query DOES contain UIFs.

For 1., 2., and 3., the answers from z3 can be kept as they are, i.e.,
should stay as EQ, EQ, and NEQ, respectively.

For 4., however, the answer should be treated as 'UNKNOWN' rather
than 'NEQ'. This is because the found counterexample might only
apply to program paths that are possible in the abstracted versions
of the programs but NOT in the concrete versions.

Note that this situation is different to the one in 2. Here, failing
to find counterexamples for the abstracted (i.e., overapproximated)
versions of the programs also implies that we cannot find any
counterexamples for the concrete versions.