Fix missing declarations in z3 queries
We can reuse most of the declarations from ARDiff.

However, for declarations such as INT_12 and REAL_34, the IDs
assigned by JPF during the ARDiff runs and our runs are different.
Because of this, we have to add the missing INT_* and REAL_*
declarations ourselves.

Furthermore, ARDiff does not provide a declaration for the 'pow'
function (it instead translates it to (^ ...) when building the
z3 query) so we have to add that one as well (alternatively, we
could also translate (pow ...) to (^ ...) in the prefix_notation
method of MathRealExpression).