Distinguish 'unknown' from 'neq but has uninterpreted functions'
Up until now, we only distinguished the following three categories:

- 'unsat' (i.e., definitely equivalent),
- 'sat' (i.e., definitely not equivalent),
- 'unknown' (i.e., cannot tell due to z3 limitations or UIF presence).

Any results with at least some degree of uncertainty (such as NEQ
results for queries that contain uninterpreted functions) were,
therefore, classified ase 'unknown'.

Because of this, it was not possible to tell during the data analysis
whether a EQ / NEQ result could not be found because of (i) z3
limitations (actual unknown results) or (ii) abstraction (i.e.,
presence of uninterpreted functions).

To enable us to better distinguish the different types of results,
we're now using the following categories:

- 'unsat (has no UIF)' (definitely equivalent),
- 'unsat (has UIF)' (definitely equivalent),
- 'sat (has no UIF)' (definitely not equivalent),
- 'sat (has UIF)' (not equivalent, but might be due to abstraction),
- 'unknown' (i.e., cannot tell due to z3 limitations).