Remove *_UIF classifications but introduce MAYBE_NEQ
UIF classifications were introduced based on the assumption that
we have to classify ALL NEQs as quasi-unknown if UIFs are present
in the z3 query.

However, as it turns out, we CAN classify some NEQ results as
'definitely NEQ' rather than 'maybe NEQ, maybe too much abstraction'.

If we get a NEQ result (i.e., sat) from a query that's looking for
models that show _non-equivalence_ of the two program versions, we
can be sure the programs are NEQ if a second query that's looking for
models that show _equivalence_ of the two program versions is unsat.

In other words: if we can show that the two programs are not just
different for some values, but ALL values, we know they are NEQ
even if UIFs are present.

Thus, knowing about UIFs doesn't really tell us much about the
classification that should be used (all EQ_UIFs should just be EQ,
and NEQ_UIFs should be NEQ if we know they are).

Because of this, the two *_UIF classifications are removed and a new
MAYBE_NEQ classification is introduced for the subset of NEQ_UIF
classifications for which do NOT know for sure whether they are NEQ.