Keep old results by default when running benchmarks
Enabling the 'clean' flag always removes all existing benchmark
results of all tools before (re-)running the benchmarks. However,
quite often, only a subset of tools is re-run, thus leading to the
loss of results for the tools that are not re-run.

Furthermore, at least with the 'diff' (rather than 'base') runner,
existing results of the specific benchmark-and-tool combination
that is scheduled to run are already removed before each run.