Add partition-level MAYBE_EQ classification
The partition-level MAYBE_EQ classification indicates that
the  programs are found to be EQ by the solver, BUT the path
condition might only be satisfiable because of the introduction
of uninterpreted functions.

Thus, for the base program without uninterpreted functions,
the correct classification of the partition might either be
EQ (if the PC is sat) or UNREACHABLE (if the PC is unsat).

---

For example, given the following program versions:

int run_v1(int x1) {
  int x2 = x1;
  if (x1 == 0) {
    return 0;
  }
  return 1;
}

int run_v2(int x1) {
  int x2 = x1;
  if (x1 == 0) {
    return 0;
  }
  return 1 + Math.abs(x2);
}

For the input 'x1 = 0', both of these programs produce the same
output of 0. For all other inputs, they produce different outputs.

The abstracted versions of the two programs look as follows:

int run_v1(int x1) {
  int x2 = UF_x2(x1);
  if (x1 == 0) {
    return 0;
  }
  return 1;
}

int run_v2(int x1) {
  int x2 = UF_x2(x1);
  if (x1 == 0) {
    return 0;
  }
  return 1 + Math.abs(x2);
}

Here, an input of 'x1 = 0' still produces the same output of 0
for both program versions.

However, the partition 'x1 = 0 & UF_x2(x1) = 0', which can be
reached by defining UF_x2 as: 'int UF_x2(int x1) { return 0; }'
also produces equivalent outputs, even though this partition
could not exist in the  base version of the programs without
uninterpreted functions.

---

Note that MAYBE_EQ on the partition-level (which indicates that
partitions are either EQ or UNREACHABLE) has different semantics
than MAYBE_EQ on the benchmark-level (which indicates that the
programs are either EQ or NEQ). This distinction has been
clarified in the description of the MAYBE_EQ enum value.