Add collection of execution traces
The collected execution traces contain information about the bytecode
instructions that are executed as part of the symbolic execution of
each input partition.

The traces are collected by the ExecutionListener and stored in
the same SQLite database file as used by the analysis scripts.

Note that the database schema used for the new data is currently
incompatible with the schema used / expected by the analysis scripts.

This incompatibility will be fixed by later commits.