Store differencing information in SQLite database
Up until now, the raw differencing information was written to text
files, interpreted by the Python analysis scripts, and then written
to the SQLite database by the Python scripts.

With this commit, the Java part of the differencing instead directly
stores the interpreted (rather than raw) differencing information
into the SQLite database (rather than text files). This makes the
data more accessible "out of the box", rather than requiring the
Python postprocessing to be run before anything useful can be done
with the data.

Note that - along with the reimplementation and refinement of the
result interpretation logic in Java - this commit also turns the
overall differencing approach into an anytime algorithm. Thus,
the differencing can now be stopped at any time and will still
produce a best-effort estimate based on the intermediate results
that have been collected up until the termination (previously,
early stopping of the algorithm would always produce a TIMEOUT
result).

The Python part of the tool is still broken at this point due to
the different database schemata that are expected on the Java vs.
the Python side of things.