Fix OutOfMemoryErrors related to large z3 query / result strings
Building the strings for the z3 queries caused OutOfMemoryErrors for
some benchmark programs, because the query strings could get quite
large for complex path conditions. Building the queries via the z3
Java API avoids this issue.

In a similar vein, printing the result strings in the differencing
driver class also caused OutOfMemoryErrors. To fix this, the results
are now simply not included in the output anymore since they aren't
currently used for the differencing anyway.