Fix base-tool error stacktraces not being written to the database
The Runner class writes stacktraces for all Exceptions that
reach it to the database. However, none of the base tool exceptions
ever reached the Runner class because all of the exceptions were
already handled by the base tool classes (SE, DSE, GradDiff,
ImpactedS).

To ensure that exceptions reach the Runner class, the base tool
classes now re-throw all exceptions after they have dealt with them.

This ensures that (i) the error messages written to the console by
the base tool classes are preserved and (ii) the Runner class now
also has the chance to write the error stacktraces to the database.