Fix overconfident NEQ classifications of ARDiff
ARDiff always returned a NEQ result when it found a NEQ example,
even if this NEQ example might only arise due to the introduction
of uninterpreted functions.

To make such results more comparable with the results of DSE, which
returns "UNKNOWN (too much abstraction)" in such cases, ARDiff now
also classifies such cases in the same way.

For comparison with the *-diff approaches, the results are treated
as MAYBE_NEQ (see: OutputClassifier).