Fix incorrect handling of partitions that throw exceptions
Up until now, partitions that throw exceptions were not dealt with
at all in the listener classes. This meant that they were neither
correctly identified as their own partitions, nor correctly checked
for equivalence across the two program versions.

This commit fixes this, thus ensuring that partitions that throw
exceptions are correctly treated as partitions and checked for
equivalence based on the type of exceptions that they throw.