Reimplement partition classification logic
The rewritten implementation makes the logic of the classification
much easier to follow and fixes several non-ideal classification
results that were identified as a result of the rewrite.

For example, if a partition contained any UIFs, it would always
be classified as MAYBE_EQ rather than EQ if there was no indication
of non-equivalence. However, ALL such cases should be classified as
EQ, since EQ can be defined as 'a partition that was fully analyzed
without finding any indication of non-equivalence'.

The new classification logic now also distinguishes between
partitions that are definitely reachable vs. maybe reachable (i.e.,
potentially only reachable due to the introduction of UIFs), whereas
the old classification logic did not distinguish these two partition
types.