Improve classification of UNKNOWN solver results
Up until now, partitions with any UNKNOWN solver results (pcStatus,
notPcStatus, eqStatus, neqStatus) were always classified as UNKNOWN.

However, if pcStatus or notPcStatus are UNKNOWN, this only reduces
the confidence of our NEQ results, forcing us to classified them
as MAYBE_NEQ rather than NEQ. EQ results are not affected at all,
because both "REACHABLE and EQ" as well as "UNREACHABLE" should both
be classified as EQ once we are at the run-level aggregation of
results.

Similarly, if eqStatus is UNKNOWN, this only reduces the confidence
of our NEQ results for partitions that are known to be REACHABLE,
forcing us to classify them as MAYBE_NEQ rather than NEQ. Results
for partitions for which we don't know for sure whether they are
reachable are not affected, because "maybe REACHABLE and NEQ" as
well as "maybe REACHABLE and MAYE_NEQ" should both be classified
as MAYBE_NEQ.

Thus, the only case in which we truly cannot make a more specific
classification than UNKNOWN is if neqStatus is UNKNOWN. Here, we
simply don't have any indication in either the EQ or NEQ direction
(similar to classifications such as DEPTH_LIMITED and TIMEOUT).