Improve classification of results with UNKNOWN neqStatus
Up until now, all partitions with an UNKNOWN neqStatus were
classified as UNKNOWN.

However, if we know that the eqStatus of a partition is SAT,
we can classify it as MAYBE_EQ instead, thereby providing
a better description of the information that we have.

Similarly, if we know that the eqStatus of a partition is UNSAT,
we can classify it as NEQ instead. This is because a partition can
only be either NEQ or EQ, so if we KNOW it is NOT EQ, it MUST be
NEQ, even if the neqStatus indicates that we're unsure about this.