Remove duplicated classification code from DifferencingListener
Up until now, the DifferencingListener contained just enough
classification code to decide whether a partition is definitely
EQ or not. This information is required to to decide whether the
return value of a areResultsEquivalent call should be true or false.

With this commit, the areResultsEquivalent decision is now delegated
to the PartitionClassifier (which already contains the necessary
classification logic anyway), thereby removing the need to maintain
(this subset of) the classification logic at two different places.