Improve classification accuracy via inclusion of hasUifPc / hasUif
If we know that a partition does not contain any UIFs, we can be
much more certain about the results of our reachability and
equivalence classifications.

For example, if both neqStatus and eqStatus are SAT (i.e., there
are both cases for which the partition is equivalent and cases for
which it isn't) we can classify this as NEQ if we know that the
partition does NOT contain any UIFs, but have to classify it as
MAYBE_NEQ if the partition DOES contain UIFs.

This reduction of certainty when classifying partitions that
contain UIFs is necessary because the introduction of UIFs leads
to an overapproximation of program behavior, and the NEQ cases
might only occur in the overapproximation, but not in the base
program without UIFs.

Therefore, knowing that some partitions do not contain UIFs can
improve overall classification accuracy for these partitions.