Keep itermediate output files for all iterations of ARDiff-base
This is mostly a preparatory change to enable us to reduce *-base
tool dependencies of ARDiff-diff in the future.

The goal here is to conduct the symbolic execution step of
ARDiff-diff iterations via DSE-diff, rather than relying
on ARDiff-base/DSE-base for all but the last iteration.

For this, we need the *.java files created by ARDiff-base at each
iteration, so we can then execute DSE-diff on them.