Keep data of all iterations in the database
Up until now, the data collected during earlier iterations was
deleted / overwritten by the data collected during later iterations.

With this commit, the data of all iterations is preserved instead.

This is accomplished through the introduction of a new `iteration`
table to the database schema that functions as a grouping mechanism
above partitions and below runs. I.e., there's a 1:n relationship
from runs to iterations and another 1:n relationship from iterations
to partitions, with each iteration containing basically the same
data as runs previously did (minus the iterationCount field).

Consequently, many of the places that previously referenced a run
object in some way now reference an iteration object instead.