Unify access to "stopwatch-times" and "split-times"
Both stopwatch-times and split-times can now be queried via
the same StopWatches.getTime(String name) method. After all,
users of the StopWatches class shouldn't have to care about the
difference between stopwatch-times and split-times.

This unified access comes with the side-effect that stopwatch-
times and split-times now share a namespace, so it is no longer
possible to create a stopwatch-time and a split-time that use
the same name.