Add resultIteration field to run model
The "result (determining) iteration" is the iteration that determines
the overall run result.

This can either be:
(a) the last iteration,
(b) the last fully analyzed iteration.

The last iteration (option (a)) is used if:
(i)   the run did not time out, or
(ii)  the run only has a single iteration, or
(iii) the result of the last iteration is a final result
      (e.g., EQ, NEQ, ... but not MAYBE_EQ, MAYBE_NEQ, ...).

Otherwise, the last fully analyzed iteration (option (b)) is used.
This is always the second-to-last iteration in such cases, because
a new iteration is only started once the previous iteration has been
fully analyzed (up to the configured depth limit).