Fix *-base tool shutdown hook not being executed sometimes
The shutdown hook was added at the start of each iteration and
removed again towards(!) the end of each iteration. This left a
gap at the end of each iteration where no shutdown hook was active.

As a result, the data for timed out runs for which the shutdown
hook was inactive at the moment of the timeout were not written
to the database.

This commit fixes this issue by ensuring that the shutdown hook
is active throughout the whole execution of the run.