Automatically create "materialized" DB views when running benchmarks
The "materialized views" have to be created at the end of a benchmark
run, i.e., once the DB is filled with all collected data.

This is because the "views" they are actually implemented as tables
created by reading data from other tables, so don't automatically
update when the data in the read-from tables changes.