Fix incomplete collection of partitionInstruction data
PartitionInstruction data was incomplete because instructions
resulting from backtracking were not counted as being part of
the corresponding partitions.

In other words, only those instructions that were explicitly
executed ("for the first time") as part of a partition were
counted as being part of a partition, whereas instructions that
were executed ("for the first time") as part of a different
partition were not.

This commit fixes this, thus ensuring that all instructions that
would be executed during a concrete execution of a program path
are counted as being part of all corresponding partitions.