Fix false positive "are_lines_reducible" classifications
This commit removes the possibility for partially analyzed runs
to be classified as having reducible lines.

The initial idea was that lines that are 100% EQ in partially
analyzed runs should be reducible. However, this turned out to
NOT hold true because we cannot be sure that the 100% EQ lines
aren't covered by any NEQ/UNKNOWN partitions which would arise
in the remaining, not-analyzed partitions.