Fix *-base tool NEQs that don't satisfy the new program summary
To understand the fixed behavior, let's look at ModDiff/Eq/Const:

```
public class oldV {
  public static double snippet(int a, int b) {
    int c=a+b;
    return c+3;
  }
}

public class newV {
  public static double snippet(int a, int b) {
    int d = 3;//change
    int c=a+b;
    return c+d;//change
  }
}
```

For this example, SE-base reports NEQ(!), even though the changes
clearly preserve the program semantics. The model that supposedly
demonstrates the non-equivalence looks as follows:

(
  (define-fun a () Int 0)
  (define-fun b () Int 0)
  (define-fun REAL_2 () Real 3.0)
  (define-fun REAL_5 () Real 4.0)
  (define-fun Ret () Real 3.0)
)

Here, REAL_2 and REAL_5 are the results of the in-to-real conversion
of the integer values `c+3` in oldV and `c+d` in newV. Thus, both
REAL_2 and REAL_5 should always be equal to `a + b + 3`, i.e.,
`0 + 0 + 3 = 3` for the given example and model.

Note that this is not the case for REAL_5, which is 4 instead of 3.

This behavior occurs because the query that is used to prove (non-)
equivalence only asserts that the specification of oldV holds, i.e.,
the query does not specify that the specification of newV should also
hold. Specifically, the query that's sent to z3 looks as follows:

```
(declare-fun ...)
...
(assert <oldVspec>)
(assert (not (= <oldVspec> <newVspec>)))
(check-sat)
(get-model)
```

Thus, the way in which the solver proves "non-equivalence" in this
case is by simply finding a model that does not satisfy newV. Of
course, this is not a valid solution in practice, because we know
that the new specification must also be satisfied (otherwise, it
wouldn't have been reported as a result by the symbolic execution).

Therefore, to fix this issues, this commit modifies the z3 queries
in such a way that the specification of newV also has to be satisfied
by NEQ model.