Fix false positive EQs for double UNKNOWN target programs
z3 reports UNSAT when checking (non-)equivalence of two UNKNOWN
programs. Since no NEQ model is provided in this case, this is
interpreted by the classification logic as an EQ result.

To fix this, we're now first checking whether the programs are
UNKNOWN. If they are, we're reporting an overall result of UNKNOWN.
If the are not, we're reporting the result of the (non-)equivalence
check.