Fix typo "&" instead of "&&" in ExecutionListener
The relevant program behavior is still the same.

The only minor difference is that && does short-circuiting, so
isIgnored() and isFirstStepInsn() are not executed anymore if they
don't need to be. However, executing them did not cause any problems
since neither of these methods actually depends on the other checked
conditions or produces any side effects.